+++ R/c2z.R
@@ -35,6 +35,39 @@ NULL
 ###############################Internal Functions###############################
 ################################################################################
 
+#' @title CheckDesc
+#' @keywords internal
+#' @noRd
+CheckDesc <- \(type, change.value = NULL) {
+
+  desc.file <- file.path(".", "DESCRIPTION")
+  desc.info <- file.info(desc.file <- file.path(".", "DESCRIPTION"))
+  desc.read <- readLines(desc.file)
+  type <- sprintf("^%s\\:", type)
+  desc.field <- grep(type, tolower(desc.read))
+
+  if (!length(desc.field)) {
+    return (NULL)
+  }
+
+  # Get current value
+  value <- Trim(gsub(".*:\\s*", "", desc.read))[[desc.field]]
+
+  # Change value if change.value is defined
+  if (!is.null(change.value)) {
+    value <- change.value
+    desc.read[desc.field] <- sprintf(
+      "%s: %s",
+      gsub(":.*", "", desc.read[desc.field]),
+      value
+    )
+    writeLines(desc.read, con = desc.file)
+  }
+
+  return (value)
+
+}
+
 #' @title FilePath
 #' @keywords internal
 #' @noRd
@@ -156,47 +189,6 @@ ListValue <- \(list, name, value) {
 
 }
 
-
-#' @title Dict
-#' @keywords internal
-#' @noRd
-Dict <- \(x = NULL, lang = "nn", data = NULL, error = NULL) {
-
-  # Visible bindings
-  key <- NULL
-
-  # Use extdata if data is not defined
-  if (is.null(data)) {
-    # Get the path to the CSV file within your package
-    csv.file <- system.file("extdata", "i18n.csv", package = "c2z")
-
-    # Read the CSV file into a data frame
-    data <- readr::read_csv(csv.file, col_types = readr::cols())
-  }
-
-  if (is.null(x)) {
-    return (data)
-  }
-
-  # Define language
-  lang <- if (!lang %in% c("nb", "nn", "no")) "en" else "no"
-
-  # Create dictionary
-  x <- data |>
-    # Filter by defined string
-    dplyr::filter(tolower(x) == tolower(key)) |>
-    # Select defined language
-    dplyr::select(!!!lang) |>
-    # Pull string
-    dplyr::pull() |>
-    # Trim
-    Trim()
-
-  if (!length(x)) x <- error
-
-  return (x)
-
-}
 
 #' @title CristinId
 #' @keywords internal
@@ -719,138 +711,6 @@ ZoteroToJson <- \(data = NULL) {
   data <- gsub("\"FALSE\"", "false", data)
   # Remove empty elements
   data <- gsub(",\"\\w+?\":\\{\\}", "", data)
-
-  return (data)
-
-}
-
-#' @title ZoteroFormat
-#' @keywords internal
-#' @noRd
-ZoteroFormat <- \(data = NULL,
-                  format = NULL,
-                  prefix = NULL) {
-
-  # Run if not empty
-  if (all(is.na(GoFish(data)))) {
-    return (NULL)
-  }
-
-  # Return as tibble if format is keys/versions
-  if (any(format %in% c("keys", "versions"))) {
-
-    # If string of characters (e.g., keys only)
-    if (format == "keys") {
-
-      data <- strsplit(data, "\n") |>
-        unlist() |>
-        (\(x) tibble::tibble(key = x))()
-
-      # Else if keys and versions
-    } else {
-
-      data <- jsonlite::fromJSON(data) |>
-        (\(x) tibble::tibble(key = names(x), version = unlist(x)))()
-
-    }
-
-    return (data)
-
-  }
-
-  # Visible bindings
-  creators <- NULL
-
-  multiline.items <- c("tags",
-                       "extra",
-                       "abstractNote",
-                       "note",
-                       "creators",
-                       "relations")
-  double.items <- c("version", "mtime")
-  list.items <- c("collections", "relations", "tags")
-
-  # Check if metadata and not in a data frame
-  if (!is.data.frame(data) &
-      (any(format == "json", is.null(format)))) {
-
-    # Check that first element of data is a list
-    if (is.list(data[[1]])) data <-  unlist(data, recursive = FALSE)
-
-    # Check all element in the meta list
-    data.list <- lapply(seq_along(data), \(i) {
-
-      # Define data
-      x <- data[[i]]
-      names <- names(data[i])
-
-      # Make certain fields not in multiline are strings
-      if (!names %in% multiline.items) x <- ToString(x)
-
-      # Add to list if element is a data frame
-      ## Make certain list.items is in a list
-      if (is.data.frame(x) | names %in% list.items) {
-        x <- if (all(is.na(GoFish(as.character(unlist(x)))))) NA else list(x)
-
-        # Make certain double.items are double
-      } else if (names %in% double.items) {
-        x <- as.double(x)
-        # Else make certain remaining items are character
-      } else {
-        x <- as.character(x)
-      }
-
-      return (x)
-
-    })
-    # Name elements
-    names(data.list) <- names(data)
-    # Keep number of columns fixed for simple conversion to tibble/JSON
-    ## Replace empty elements with NA
-    data.list[lengths(data.list) == 0] <- NA
-    # Set key and initial version is missing
-    if (!"key" %in% names(data.list)) {
-      data.list <- c(key = ZoteroKey(), version = 0, data.list)
-    }
-    # Remove elements not in category if item
-    if (!"parentCollection" %in% names(data.list)) {
-      data.list <- data.list[
-        names(data.list) %in%
-          c("key", "version", ZoteroTypes(data.list$itemType))]
-    }
-
-    # Format as tibble and remove empty elements
-    data <- tibble::as_tibble(data.list[lengths(data.list) != 0])
-
-  }
-
-  # Set data as tibble if data frame
-  if (is.data.frame(data)) {
-
-    data <- tibble::as_tibble(data) |>
-      # Replace empty string with NA
-      dplyr::mutate_if(is.character, list(~dplyr::na_if(., ""))) |>
-      dplyr::mutate(
-        # Make certain tags is a data.frame within a list
-        dplyr::across(
-          dplyr::any_of("tags"), ~ purrr::map(tags, ~ {
-            if (all(!is.na(.x))) as.data.frame(.x)
-          })
-        ),
-        # Make certain that parentCollection is a character
-        dplyr::across(dplyr::any_of("parentCollection"), as.character),
-        # Add prefix if defined
-        prefix = GoFish(prefix),
-        # Fix creators
-        creators = GoFish(purrr::map(creators, FixCreators))
-      ) |>
-      # Remove empty columns
-      dplyr::select(dplyr::where(~sum(!is.na(.x)) > 0))
-
-    # Else convert to string
-  } else {
-    data <- ToString(data, "\n")
-  }
 
   return (data)